Add files via upload
there are two files; an html file and a script file, the map will open once the file is pressed, using geolocation and permission from user the map will set up on where it is being opened from ( current coordinates ) with an error location if device location is unavailable, being Amman,jordan. On the upper left corner there are navigation buttons to zoom in and out and freely rotate, in addition to a search bar that helps set up the map wherever the user enters.